Migrate Marhsalled sessions to JSON
**Why**: Since we already have sessions in production that were
serialized with Marshal, we need to be able to continue supporting them,
and migrate them to the new JSON serializer. By logging the presence of
Marshalled sessions, we can check the logs after a predefined amount of
time to determine when it's safe to remove this migration code.

The first 3 characters of the Marshal-serialized + Base64 encoded value
are always the same as far as I can tell. I tested this locally and on
the dev and qa servers. Since these are constant, we can check if the
decrypted value passed into `SessionEncryptor.load` starts with those 3
characters, and if so, it means it was serialized with Marshal. We can
then serialize it with JSON. Eventually, all sessions that were
serialized with Marshal will be converted to JSON.