Memoize MARSHAL_SIGNATURE constant
**Why**: Due to Rails app reloading, the constant gets defined twice
because the `SessionEncryptor` class gets called from an initializer.
This outputs a warning to the Rails log when running the app:
`already initialized constant SessionEncryptor::MARSHAL_SIGNATURE`.

One way to get rid of the warning is with memoization, but by doing
that, we run into a Rubocop bug:
https://github.com/bbatsov/rubocop/issues/4249

While we wait for the Rubocop fix to be available in a new gem release,
I've disabled the cop that is failing.

Alternatively, we could do `unless const_defined?(:MARSHAL_SIGNATURE)`,
but I thought memoization was cleaner.